requirements.txt
streamlit==1.22.0
opencv-python==4.8.1.78
numpy==1.26.3
Pillow==10.2.0
pandas==2.2.2